[CCD-91] Change submission id from int to long

+++ src/main/java/github/clone_code_detection/entity/moodle/DetectRequest.java
@@ -14,5 +14,5 @@ import java.util.List;
 @AllArgsConstructor
 public class DetectRequest {
     @JsonProperty("submission_ids")
-    List<Integer> submissionIds;
+    List<Long> submissionIds;
 }
